Introduce individual Event class and refactor EventDetection to use it
- Rename legacy `event.py` to `detection.py`.
- Create a new `event.py` with an `Event` class that represents individual events and all of their properties, including an `excluded` boolean flag.
- Refactor `EventDetection` and `EventAnalysis` to use a list of `Event` objects (`self.detected_events`), replacing the numerous separate parallel arrays while maintaining full backward-compatibility using `@property` getters and setters.
- Add comprehensive test coverage in `tests/test_event.py` for both the `Event` class and the array property mappings.
- Fix legacy test parameter alignment inside `tests/test_miniML_functions.py` and `tests/test_miniML_updated_functions.py`.
