+++ src/miniml/__init__.py
@@ -1,4 +1,5 @@
-from miniml.core.event import EventAnalysis, EventDetection, EventStats
+from miniml.core.detection import EventAnalysis, EventDetection, EventStats
+from miniml.core.event import Event
 from miniml.core.trace import MiniTrace
 from miniml.settings import Settings
 from miniml.fileio.trace_loader import TraceLoader
